+++ pollis-core/src/commands/mls/group_state.rs
@@ -1,0 +1,698 @@
+//! MLS group lifecycle: create / load / forget / encrypt / decrypt /
+//! commit processing / external join / GroupInfo publishing.
+
+use openmls::prelude::group_info::VerifiableGroupInfo;
+use openmls::prelude::*;
+use openmls_basic_credential::SignatureKeyPair;
+use openmls_traits::OpenMlsProvider;
+
+use std::sync::Arc;
+use tls_codec::{Deserialize as TlsDeserialize, Serialize as TlsSerialize};
+
+use crate::error::Result;
+use crate::state::AppState;
+
+use super::device::{load_or_create_device_signer, verify_added_devices};
+use super::provider::{make_credential, PollisProvider, CS};
+
+// ── GroupInfo publishing ─────────────────────────────────────────────────────
+
+/// Export a fresh `GroupInfo` for the given conversation and upsert it
+/// into the remote `mls_group_info` table. Called by every device that
+/// merges a commit (the originator right after `merge_pending_commit`,
+/// receivers right after `merge_staged_commit`).
+///
+/// The row is conversation-scoped and only overwritten with a STRICTLY
+/// greater epoch, so concurrent writers at the same epoch are idempotent
+/// and receivers don't waste work once the committer has already
+/// published.
+///
+/// No-op if:
+///   - the device has no local MLS group for this conversation
+///   - the device has no `account_id_key` (pre-enrollment)
+///
+/// This function is the prerequisite for the Secret Key recovery path:
+/// a brand-new device uses the stored `GroupInfo` to construct an MLS
+/// external commit joining the group, without needing a Welcome.
+pub async fn publish_group_info(
+    state: &Arc<AppState>,
+    conversation_id: &str,
+) -> crate::error::Result<()> {
+    // Sync scope: load the local group, recover the signer, export a
+    // GroupInfo, and TLS-serialize it. Nothing !Send crosses await.
+    let device_id_opt = state.device_id.lock().await.clone();
+    let Some(device_id) = device_id_opt else {
+        return Ok(());
+    };
+
+    let exported: Option<(u64, Vec<u8>)> = {
+        let guard = state.local_db.lock().await;
+        let Some(db) = guard.as_ref() else {
+            return Ok(());
+        };
+        let provider = PollisProvider::new(db.conn());
+        let (group, signer) = match load_group_with_signer(&provider, conversation_id) {
+            Ok(pair) => pair,
+            Err(_) => return Ok(()),
+        };
+        let epoch = group.epoch().as_u64();
+        let msg = match group.export_group_info(provider.crypto(), &signer, true) {
+            Ok(m) => m,
+            Err(e) => {
+                eprintln!("[mls] publish_group_info: export failed for {conversation_id}: {e}");
+                return Ok(());
+            }
+        };
+        let bytes = msg
+            .tls_serialize_detached()
+            .map_err(|e| crate::error::Error::Other(anyhow::anyhow!("group_info serialize: {e}")))?;
+        Some((epoch, bytes))
+    };
+
+    let Some((epoch, bytes)) = exported else {
+        return Ok(());
+    };
+
+    let conn = state.remote_db.conn().await?;
+    conn.execute(
+        "INSERT INTO mls_group_info \
+         (conversation_id, epoch, group_info, updated_at, updated_by_device_id) \
+         VALUES (?1, ?2, ?3, datetime('now'), ?4) \
+         ON CONFLICT(conversation_id) DO UPDATE SET \
+             epoch = excluded.epoch, \
+             group_info = excluded.group_info, \
+             updated_at = datetime('now'), \
+             updated_by_device_id = excluded.updated_by_device_id \
+         WHERE excluded.epoch > mls_group_info.epoch",
+        libsql::params![conversation_id, epoch as i64, bytes, device_id],
+    )
+    .await?;
+
+    Ok(())
+}
+
+// ── External-commit joining ──────────────────────────────────────────────────
+
+/// Join an existing MLS group via external commit, using the latest
+/// `GroupInfo` blob stored server-side in `mls_group_info`. The new
+/// device becomes a full member of the group at the epoch *after* the
+/// one carried in the GroupInfo.
+///
+/// Used by the Secret Key recovery path: when a new device recovers
+/// `account_id_key` without any sibling device online to issue a
+/// Welcome, it fetches each of the user's groups' GroupInfo and
+/// externally commits into them. The commit is posted to
+/// `mls_commit_log` so existing members will merge it on their next
+/// `process_pending_commits` pass.
+///
+/// Safety note: this path does NOT currently pass through the outbound
+/// cross-signing cert check. Existing members that implement the
+/// step-3b inbound cert verification will reject external-join commits
+/// from devices whose cert doesn't chain to the user's
+/// `account_id_pub` — which is exactly the desired behavior.
+pub async fn external_join_group(
+    state: &Arc<AppState>,
+    conversation_id: &str,
+    user_id: &str,
+) -> crate::error::Result<()> {
+    let device_id = state
+        .device_id
+        .lock()
+        .await
+        .clone()
+        .ok_or_else(|| crate::error::Error::Other(anyhow::anyhow!("device_id not set")))?;
+
+    // 1. Fetch the stored GroupInfo for this conversation.
+    let (group_info_bytes, stored_epoch): (Vec<u8>, i64) = {
+        let conn = state.remote_db.conn().await?;
+        let mut rows = conn
+            .query(
+                "SELECT group_info, epoch FROM mls_group_info WHERE conversation_id = ?1",
+                libsql::params![conversation_id],
+            )
+            .await?;
+        match rows.next().await? {
+            Some(row) => (row.get(0)?, row.get(1)?),
+            None => {
+                return Err(crate::error::Error::Other(anyhow::anyhow!(
+                    "no GroupInfo stored for {conversation_id} — cannot external-join"
+                )))
+            }
+        }
+    };
+
+    // 2. Run the external commit inside the local_db sync scope.
+    let commit_bytes: Vec<u8> = {
+        let guard = state.local_db.lock().await;
+        let db = guard.as_ref().ok_or_else(|| {
+            crate::error::Error::Other(anyhow::anyhow!("Not signed in"))
+        })?;
+        let provider = PollisProvider::new(db.conn());
+
+        let mut env_reader: &[u8] = &group_info_bytes;
+        let msg_in = MlsMessageIn::tls_deserialize(&mut env_reader).map_err(|e| {
+            crate::error::Error::Other(anyhow::anyhow!(
+                "stored group_info envelope failed to deserialize: {e}"
+            ))
+        })?;
+        let verifiable_group_info: VerifiableGroupInfo = match msg_in.extract() {
+            MlsMessageBodyIn::GroupInfo(gi) => gi,
+            other => {
+                return Err(crate::error::Error::Other(anyhow::anyhow!(
+                    "expected GroupInfo in mls_group_info, got {:?}",
+                    std::mem::discriminant(&other)
+                )));
+            }
+        };
+
+        // Load (or create) this device's stable MLS signing keypair.
+        let (sig_keys, sig_pub_bytes) =
+            load_or_create_device_signer(&provider, user_id, &device_id)?;
+
+        let credential = make_credential(user_id, &device_id);
+        let sig_pub = OpenMlsSignaturePublicKey::new(
+            sig_pub_bytes.into(),
+            CS.signature_algorithm(),
+        )
+        .map_err(|e| crate::error::Error::Other(anyhow::anyhow!("sig pub key: {e}")))?;
+        let cred_with_key = CredentialWithKey {
+            credential,
+            signature_key: sig_pub.into(),
+        };
+
+        // Drop any stale local group with the same ID so the external
+        // commit builder doesn't collide.
+        let group_id = GroupId::from_slice(conversation_id.as_bytes());
+        if let Ok(Some(mut old)) = MlsGroup::load(provider.storage(), &group_id) {
+            let _ = old.delete(provider.storage());
+        }
+
+        let join_config = MlsGroupJoinConfig::builder()
+            .use_ratchet_tree_extension(true)
+            .build();
+
+        let (_joined_group, commit_bundle) = MlsGroup::external_commit_builder()
+            .with_config(join_config)
+            .build_group(&provider, verifiable_group_info, cred_with_key)
+            .map_err(|e| crate::error::Error::Other(anyhow::anyhow!(
+                "external commit build_group: {e}"
+            )))?
+            .load_psks(provider.storage())
+            .map_err(|e| crate::error::Error::Other(anyhow::anyhow!(
+                "external commit load_psks: {e}"
+            )))?
+            .build(provider.rand(), provider.crypto(), &sig_keys, |_| true)
+            .map_err(|e| crate::error::Error::Other(anyhow::anyhow!(
+                "external commit build: {e}"
+            )))?
+            .finalize(&provider)
+            .map_err(|e| crate::error::Error::Other(anyhow::anyhow!(
+                "external commit finalize: {e}"
+            )))?;
+
+        let (commit_msg, _welcome_msg, _new_group_info) = commit_bundle.into_contents();
+        commit_msg
+            .tls_serialize_detached()
+            .map_err(|e| crate::error::Error::Other(anyhow::anyhow!("commit serialize: {e}")))?
+    };
+
+    // 3. Post the commit to mls_commit_log so existing members will
+    //    process it on their next process_pending_commits pass.
+    let conn = state.remote_db.conn().await?;
+    conn.execute(
+        "INSERT INTO mls_commit_log \
+         (conversation_id, epoch, sender_id, commit_data, added_user_id, added_device_ids) \
+         VALUES (?1, ?2, ?3, ?4, ?5, ?6)",
+        libsql::params![
+            conversation_id,
+            stored_epoch,
+            user_id,
+            commit_bytes,
+            user_id,
+            device_id.clone()
+        ],
+    )
+    .await?;
+
+    // 4. Refresh the stored GroupInfo at the new epoch so any NEXT
+    //    new device joining via this same path sees the up-to-date
+    //    tree.
+    if let Err(e) = publish_group_info(state, conversation_id).await {
+        eprintln!(
+            "[mls] external_join_group: publish_group_info failed (non-fatal): {e}"
+        );
+    }
+
+    eprintln!(
+        "[mls] external_join_group: {user_id}:{device_id} joined {conversation_id} from epoch {stored_epoch}"
+    );
+
+    Ok(())
+}
+
+// ── Phase 3: Group / DM creation ─────────────────────────────────────────────
+
+/// Internal: create a fresh MLS group for `conversation_id` with
+/// `creator_user_id` as the sole initial member.  Group state is persisted in
+/// the local `mls_kv` table via `MlsStore`.
+///
+/// `use_ratchet_tree_extension(true)` is set so that Welcome messages sent in
+/// Phase 4 embed the full ratchet tree — recipients can join without a separate
+/// out-of-band tree download.
+pub async fn init_mls_group(
+    state: &Arc<AppState>,
+    conversation_id: &str,
+    creator_user_id: &str,
+) -> Result<()> {
+    let device_id = state.device_id.lock().await.clone()
+        .ok_or_else(|| crate::error::Error::Other(anyhow::anyhow!("device_id not set")))?;
+
+    // Scope the local_db guard so it is dropped before the async
+    // publish_group_info call below (which re-acquires it).
+    {
+        let guard = state.local_db.lock().await;
+        let db = guard.as_ref().ok_or_else(|| {
+            crate::error::Error::Other(anyhow::anyhow!("Not signed in"))
+        })?;
+        let provider = PollisProvider::new(db.conn());
+
+        let (sig_keys, sig_pub_bytes) =
+            load_or_create_device_signer(&provider, creator_user_id, &device_id)?;
+
+        let credential = make_credential(creator_user_id, &device_id);
+        let sig_pub = OpenMlsSignaturePublicKey::new(
+            sig_pub_bytes.into(),
+            CS.signature_algorithm(),
+        ).map_err(|e| crate::error::Error::Other(anyhow::anyhow!("sig pub key: {e}")))?;
+        let cred_with_key = CredentialWithKey {
+            credential,
+            signature_key: sig_pub.into(),
+        };
+
+        let group_id = GroupId::from_slice(conversation_id.as_bytes());
+
+        // Delete any stale group with the same ID so the create below never
+        // collides.  This is a no-op on first creation and essential during
+        // repair (where the old group still exists but is broken/outdated).
+        if let Ok(Some(mut old)) = MlsGroup::load(provider.storage(), &group_id) {
+            let _ = old.delete(provider.storage());
+        }
+
+        let config = MlsGroupCreateConfig::builder()
+            .ciphersuite(CS)
+            .use_ratchet_tree_extension(true)
+            .build();
+
+        MlsGroup::new_with_group_id(&provider, &sig_keys, &config, group_id, cred_with_key)
+            .map_err(|e| crate::error::Error::Other(anyhow::anyhow!("create mls group: {e}")))?;
+    }
+
+    // Publish the epoch-0 GroupInfo so a future device enrolling via the
+    // Secret Key path can join this group via external commit.
+    if let Err(e) = publish_group_info(state, conversation_id).await {
+        eprintln!("[mls] init_mls_group: publish_group_info failed (non-fatal): {e}");
+    }
+
+    Ok(())
+}
+
+/// Create a fresh MLS group for `conversation_id` (a channel or DM ULID).
+/// The creator becomes the sole initial member.  Other users are added via
+/// `reconcile_group_mls`.
+pub async fn create_mls_group(
+    state: &Arc<AppState>,
+    conversation_id: String,
+    creator_user_id: String,
+) -> Result<()> {
+    init_mls_group(state, &conversation_id, &creator_user_id).await
+}
+
+// ── Phase 4: Member changes ───────────────────────────────────────────────────
+
+/// Reload an existing MLS group from storage and recover the signer.
+///
+/// Returns `(MlsGroup, SignatureKeyPair)` ready for use with the provider whose
+/// connection was passed to `PollisProvider::new`.
+pub(super) fn load_group_with_signer(
+    provider: &PollisProvider<'_>,
+    conversation_id: &str,
+) -> crate::error::Result<(MlsGroup, SignatureKeyPair)> {
+    let group_id = GroupId::from_slice(conversation_id.as_bytes());
+
+    let mut group = MlsGroup::load(provider.storage(), &group_id)
+        .map_err(|e| crate::error::Error::Other(anyhow::anyhow!("mls load: {e}")))?
+        .ok_or_else(|| crate::error::Error::Other(anyhow::anyhow!(
+            "MLS group not found for conversation {conversation_id}"
+        )))?;
+
+    // Retrieve the signature public key stored in the group's leaf node, then
+    // read back the full keypair from mls_kv.
+    let sig_pub_bytes = group
+        .own_leaf_node()
+        .ok_or_else(|| crate::error::Error::Other(anyhow::anyhow!("no own leaf node")))?
+        .signature_key()
+        .as_slice()
+        .to_vec();
+
+    let signer = SignatureKeyPair::read(
+        provider.storage(),
+        &sig_pub_bytes,
+        CS.signature_algorithm(),
+    )
+    .ok_or_else(|| crate::error::Error::Other(anyhow::anyhow!("signer not found in mls_kv")))?;
+
+    // Resolve any in-flight pending commit so the group is operational before
+    // the caller performs new operations.
+    group
+        .merge_pending_commit(provider)
+        .map_err(|e| crate::error::Error::Other(anyhow::anyhow!("merge pending: {e}")))?;
+
+    Ok((group, signer))
+}
+
+/// Wipe all local MLS state for a group without publishing a commit.
+///
+/// Used when the local user leaves a group.  MLS does not allow a member to
+/// commit their own removal (`remove_members` with self as target errors), so
+/// instead we just delete the local group epoch.  The remaining members still
+/// have this user in their group state until the next admin-issued commit, but
+/// forward secrecy ensures the leaver cannot decrypt messages after the next
+/// epoch advance.
+pub async fn forget_local_mls_group(
+    state: &Arc<AppState>,
+    group_id: &str,
+) -> crate::error::Result<()> {
+    let guard = state.local_db.lock().await;
+    let db = guard.as_ref().ok_or_else(|| {
+        crate::error::Error::Other(anyhow::anyhow!("Not signed in"))
+    })?;
+    let provider = PollisProvider::new(db.conn());
+    let mls_group_id = GroupId::from_slice(group_id.as_bytes());
+
+    if let Ok(Some(mut group)) = MlsGroup::load(provider.storage(), &mls_group_id) {
+        group.delete(provider.storage())
+            .map_err(|e| crate::error::Error::Other(anyhow::anyhow!("mls delete group: {e}")))?;
+    }
+    // If the group wasn't found locally, nothing to clean up.
+    Ok(())
+}
+
+/// Apply any commits from `mls_commit_log` that this member has not yet seen.
+///
+/// Reads rows where `epoch >= current_local_epoch` in ascending order, applies
+/// each commit, and advances the local epoch.  An epoch gap (unexpected jump)
+/// stops processing and logs an error — this indicates a missed or reordered
+/// commit that would require manual intervention in a production system.
+///
+/// `mls_group_id` must already be resolved (group_id for channels,
+/// conversation_id for DMs).
+/// Ensure this device has a local MLS group at the latest epoch for
+/// `mls_group_id`. Processes any pending commits from the commit log.
+/// If the local group is missing, evicted, or unrecoverably behind,
+/// falls back to external-join using the published GroupInfo.
+///
+/// `user_id` is needed for the external-join fallback.
+pub async fn process_pending_commits_inner(
+    state: &Arc<AppState>,
+    mls_group_id: &str,
+    user_id: &str,
+) -> crate::error::Result<()> {
+    // 1. Get the current epoch from the local group.
+    let has_group = {
+        let guard = state.local_db.lock().await;
+        let db = guard.as_ref().ok_or_else(|| {
+            crate::error::Error::Other(anyhow::anyhow!("Not signed in"))
+        })?;
+        let provider = PollisProvider::new(db.conn());
+        let group_id = GroupId::from_slice(mls_group_id.as_bytes());
+        MlsGroup::load(provider.storage(), &group_id)
+            .ok()
+            .flatten()
+            .map(|g| g.epoch().as_u64())
+    };
+    let initial_epoch = match has_group {
+        Some(epoch) => epoch,
+        None => {
+            // No local group — external-join to create one.
+            if let Err(e) = external_join_group(state, mls_group_id, user_id).await {
+                eprintln!("[mls] process_pending_commits: no local group for {mls_group_id}, external-join failed: {e}");
+            }
+            return Ok(());
+        }
+    };
+
+    // 2. Fetch pending commits from remote, along with the add-metadata
+    //    columns (`added_user_id`, `added_device_ids`) so we can verify
+    //    cross-signing certs BEFORE calling `process_message`. Collected
+    //    into an owned Vec so the `rows` cursor is dropped before any
+    //    local-DB await below.
+    let conn = state.remote_db.conn().await?;
+    let mut rows = conn.query(
+        "SELECT epoch, commit_data, added_user_id, added_device_ids \
+         FROM mls_commit_log \
+         WHERE conversation_id = ?1 AND epoch >= ?2 \
+         ORDER BY epoch ASC, seq ASC",
+        libsql::params![mls_group_id, initial_epoch as i64],
+    ).await?;
+
+    #[derive(Debug)]
+    struct PendingCommit {
+        epoch: i64,
+        commit_data: Vec<u8>,
+        added_user_id: Option<String>,
+        added_device_ids: Vec<String>,
+    }
+
+    let mut pending: Vec<PendingCommit> = Vec::new();
+    while let Some(row) = rows.next().await? {
+        let epoch: i64 = row.get(0)?;
+        let data: Vec<u8> = row.get(1)?;
+        let added_user_id: Option<String> = row.get::<Option<String>>(2).ok().flatten();
+        let ids_csv: Option<String> = row.get::<Option<String>>(3).ok().flatten();
+        let added_device_ids: Vec<String> = ids_csv
+            .as_deref()
+            .map(|s| {
+                s.split(',')
+                    .map(|x| x.trim().to_string())
+                    .filter(|x| !x.is_empty())
+                    .collect()
+            })
+            .unwrap_or_default();
+        pending.push(PendingCommit {
+            epoch,
+            commit_data: data,
+            added_user_id,
+            added_device_ids,
+        });
+    }
+    drop(rows);
+
+    // 3. Apply each commit in epoch order. For any commit carrying add
+    //    metadata, verify every added device's cross-signing cert
+    //    against the user's account_id_pub BEFORE touching the group
+    //    state.
+    let mut current_epoch = initial_epoch;
+    let mut any_applied = false;
+    'commit_loop: for commit in pending {
+        if commit.epoch as u64 != current_epoch {
+            eprintln!(
+                "[mls] process_pending_commits: epoch gap for {mls_group_id}: \
+                 expected {current_epoch}, got {} — stopping",
+                commit.epoch
+            );
+            break;
+        }
+
+        // ── Inbound cert verification (advisory) ─────────────────
+        // Log a warning if cross-signing verification fails, but still
+        // process the commit. Blocking here causes epoch divergence
+        // because the commit was already applied by the sender.
+        if let Some(ref added_user_id) = commit.added_user_id {
+            let ok = verify_added_devices(
+                &conn,
+                added_user_id,
+                &commit.added_device_ids,
+            )
+            .await;
+            match ok {
+                Ok(true) => {}
+                Ok(false) => {
+                    eprintln!(
+                        "[mls] process_pending_commits: WARN cross-signing verification failed for {added_user_id} at epoch {} in {mls_group_id} — processing anyway",
+                        commit.epoch
+                    );
+                }
+                Err(e) => {
+                    eprintln!(
+                        "[mls] process_pending_commits: WARN cert verification error for {mls_group_id}: {e} — processing anyway"
+                    );
+                }
+            }
+        }
+
+        let commit_data = commit.commit_data;
+
+        // All MLS work is synchronous and scoped so nothing !Send crosses
+        // the lock().await boundary.
+        let applied = {
+            let guard = state.local_db.lock().await;
+            let db = match guard.as_ref() {
+                Some(db) => db,
+                None => break,
+            };
+            let provider = PollisProvider::new(db.conn());
+            let group_id = GroupId::from_slice(mls_group_id.as_bytes());
+            let mut group = match MlsGroup::load(provider.storage(), &group_id)
+                .map_err(|e| crate::error::Error::Other(anyhow::anyhow!("mls load: {e}")))?
+            {
+                Some(g) => g,
+                None => break,
+            };
+
+            let mut reader: &[u8] = &commit_data;
+            let msg_in = match MlsMessageIn::tls_deserialize(&mut reader) {
+                Ok(m) => m,
+                Err(e) => {
+                    eprintln!("[mls] process_pending_commits: deserialize failed for {mls_group_id} at epoch {}: {e} — stopping", commit.epoch);
+                    break;
+                }
+            };
+            let protocol_msg = match msg_in.try_into_protocol_message() {
+                Ok(m) => m,
+                Err(e) => {
+                    eprintln!("[mls] process_pending_commits: protocol msg failed for {mls_group_id} at epoch {}: {e} — stopping", commit.epoch);
+                    break;
+                }
+            };
+
+            match group.process_message(&provider, protocol_msg) {
+                Ok(processed) => {
+                    if let ProcessedMessageContent::StagedCommitMessage(staged) = processed.into_content() {
+                        if let Err(e) = group.merge_staged_commit(&provider, *staged) {
+                            eprintln!("[mls] process_pending_commits: merge failed for {mls_group_id} at epoch {}: {e} — stopping", commit.epoch);
+                            break;
+                        }
+                    }
+                }
+                Err(e) => {
+                    let msg = format!("{e}");
+                    // If we were evicted (kicked), delete the stale group so
+                    // external-join recovery can create a fresh one.
+                    if msg.contains("evicted") {
+                        eprintln!("[mls] process_pending_commits: evicted from {mls_group_id} — deleting local group for recovery");
+                        let _ = group.delete(provider.storage());
+                    } else {
+                        eprintln!("[mls] process_pending_commits: {e} for {mls_group_id} at epoch {} — stopping", commit.epoch);
+                    }
+                    break;
+                }
+            }
+
+            true
+        };
+
+        if applied {
+            current_epoch += 1;
+            any_applied = true;
+        }
+    }
+
+    if any_applied {
+        if let Err(e) = publish_group_info(state, mls_group_id).await {
+            eprintln!("[mls] process_pending_commits: publish_group_info failed (non-fatal): {e}");
+        }
+        // Voice E2EE: when the epoch advances for the MLS group currently
+        // backing the active voice room, re-derive the per-room key and
+        // rotate it on the live KeyProvider. No-op when voice is idle.
+        crate::commands::voice_e2ee::on_mls_epoch_changed(state, mls_group_id).await;
+    }
+
+    // If the group was deleted during processing (e.g. eviction),
+    // external-join to recover.
+    let group_exists = {
+        let guard = state.local_db.lock().await;
+        guard.as_ref().map_or(false, |db| {
+            has_local_group(db.conn(), mls_group_id)
+        })
+    };
+    if !group_exists {
+        eprintln!("[mls] process_pending_commits: group {mls_group_id} was deleted during processing — external-joining to recover");
+        if let Err(e) = external_join_group(state, mls_group_id, user_id).await {
+            eprintln!("[mls] process_pending_commits: recovery external-join failed for {mls_group_id}: {e}");
+        }
+    }
+
+    Ok(())
+}
+
+/// Tauri command wrapper — resolves conversation_id to MLS group ID, then
+/// delegates to `process_pending_commits_inner`.
+pub async fn process_pending_commits(
+    state: &Arc<AppState>,
+    conversation_id: String,
+    user_id: String,
+) -> crate::error::Result<()> {
+    let mls_group_id = {
+        let conn = state.remote_db.conn().await?;
+        let mut rows = conn.query(
+            "SELECT group_id FROM channels WHERE id = ?1",
+            libsql::params![conversation_id.clone()],
+        ).await?;
+        match rows.next().await? {
+            Some(row) => row.get::<String>(0)?,
+            None => conversation_id,
+        }
+    };
+    process_pending_commits_inner(state, &mls_group_id, &user_id).await
+}
+
+// ── Phase 5 helpers: encrypt / decrypt ───────────────────────────────────────
+
+/// Check whether an MLS group exists in the local database.
+pub fn has_local_group(conn: &rusqlite::Connection, conversation_id: &str) -> bool {
+    let provider = PollisProvider::new(conn);
+    let group_id = GroupId::from_slice(conversation_id.as_bytes());
+    matches!(MlsGroup::load(provider.storage(), &group_id), Ok(Some(_)))
+}
+
+/// Try to encrypt `plaintext` with the MLS group for `conversation_id`.
+///
+/// Returns `None` — without logging — if the group does not exist locally
+/// (e.g. the channel was created before MLS was rolled out).  The caller
+/// should fall back to the legacy Signal sender-key path in that case.
+pub fn try_mls_encrypt(
+    conn: &rusqlite::Connection,
+    conversation_id: &str,
+    plaintext: &[u8],
+) -> Option<Vec<u8>> {
+    let provider = PollisProvider::new(conn);
+    let (mut group, signer) = load_group_with_signer(&provider, conversation_id).ok()?;
+    let msg_out = group.create_message(&provider, &signer, plaintext).ok()?;
+    msg_out.tls_serialize_detached().ok()
+}
+
+/// Try to decrypt MLS ciphertext bytes for `conversation_id`.
+///
+/// The bytes must be TLS-serialised `MlsMessageOut` (i.e. what we stored in
+/// `message_envelope.ciphertext` after `send_message` used MLS).  Returns
+/// the raw plaintext bytes on success, or `None` if the bytes are not a
+/// valid MLS `ApplicationMessage` or if decryption fails for any reason.
+pub fn try_mls_decrypt(
+    conn: &rusqlite::Connection,
+    conversation_id: &str,
+    ciphertext: &[u8],
+) -> Option<Vec<u8>> {
+    let provider = PollisProvider::new(conn);
+    let group_id = GroupId::from_slice(conversation_id.as_bytes());
+    let mut group = MlsGroup::load(provider.storage(), &group_id).ok()??;
+
+    let mut reader: &[u8] = ciphertext;
+    let msg_in = MlsMessageIn::tls_deserialize(&mut reader).ok()?;
+    let protocol_msg = msg_in.try_into_protocol_message().ok()?;
+    let processed = group.process_message(&provider, protocol_msg).ok()?;
+
+    match processed.into_content() {
+        ProcessedMessageContent::ApplicationMessage(app_msg) => Some(app_msg.into_bytes()),
+        _ => None,
+    }
+}
