+++ pollis-core/src/commands/mls/mod.rs
@@ -1,0 +1,44 @@
+//! MLS commands — split into cohesive submodules. Public surface is
+//! preserved via the `pub use` re-exports below so every external caller
+//! (Tauri shims, sibling `commands::*` modules, integration tests) keeps
+//! resolving names at `pollis_core::commands::mls::*`.
+
+mod device;
+mod group_state;
+mod key_packages;
+mod provider;
+mod reconcile;
+mod welcomes;
+
+// ── Provider / credential helpers ────────────────────────────────────────────
+pub use provider::{
+    make_credential, parse_credential_device_id, parse_credential_user_id, PollisProvider,
+};
+
+// ── Per-device signing keys + cross-signing ──────────────────────────────────
+pub use device::{ensure_device_cert, load_or_create_device_signer, resign_stale_device_certs};
+
+// ── Key packages ─────────────────────────────────────────────────────────────
+pub use key_packages::{
+    ensure_mls_key_package, fetch_mls_key_package, generate_mls_key_package,
+    publish_mls_key_package, validate_key_package,
+};
+
+// ── Welcomes ─────────────────────────────────────────────────────────────────
+pub use welcomes::{apply_welcome, poll_mls_welcomes, poll_mls_welcomes_inner, process_welcome};
+
+// ── Group lifecycle / encrypt / decrypt / commit processing ──────────────────
+pub use group_state::{
+    create_mls_group, external_join_group, forget_local_mls_group, has_local_group,
+    init_mls_group, process_pending_commits, process_pending_commits_inner, publish_group_info,
+    try_mls_decrypt, try_mls_encrypt,
+};
+
+// ── Reconcile + self-repair ──────────────────────────────────────────────────
+pub use reconcile::{
+    reconcile_group_mls, reconcile_group_mls_core, reconcile_group_mls_core_staged,
+    reconcile_group_mls_impl, repair_mls_group, ReconcileCommitData, ReconcileOutcome,
+};
+
+#[cfg(test)]
+mod tests;
